+++ cli/cmd/init.go
@@ -37,7 +37,7 @@ func makeHappyClient(cmd *cobra.Command, sliceName, stackName string, tags []str
 		return nil, err
 	}
 	ctx := cmd.Context()
-	awsBackend, err := backend.NewAWSBackend(ctx, happyConfig)
+	awsBackend, err := backend.NewAWSBackend(ctx, happyConfig.GetEnvironmentContext())
 	if err != nil {
 		return nil, err
 	}
@@ -54,6 +54,7 @@ func makeHappyClient(cmd *cobra.Command, sliceName, stackName string, tags []str
 	}
 	workspaceRepo := createWorkspaceRepo(dryRun, awsBackend)
 	stackService := stackservice.NewStackService().
+		WithHappyConfig(happyConfig).
 		WithBackend(awsBackend).
 		WithWorkspaceRepo(workspaceRepo)
 
@@ -84,6 +85,7 @@ func configureArtifactBuilder(
 	happyConfig *config.HappyConfig,
 	backend *backend.Backend) (ab.ArtifactBuilderIface, map[string]string, error) {
 	artifactBuilder := ab.NewArtifactBuilder(dryRun).
+		WithHappyConfig(happyConfig).
 		WithConfig(builderConfig).
 		WithBackend(backend)
 	var err error
@@ -188,12 +190,12 @@ func validate(validations ...validation) error {
 	return nil
 }
 
-func makeWaitOptions(stackName string, backend *backend.Backend) waitoptions.WaitOptions {
-	taskOrchestrator := orchestrator.NewOrchestrator().WithBackend(backend)
+func makeWaitOptions(stackName string, happyConfig *config.HappyConfig, backend *backend.Backend) waitoptions.WaitOptions {
+	taskOrchestrator := orchestrator.NewOrchestrator().WithHappyConfig(happyConfig).WithBackend(backend)
 	return waitoptions.WaitOptions{
 		StackName:    stackName,
 		Orchestrator: taskOrchestrator,
-		Services:     backend.Conf().GetServices(),
+		Services:     happyConfig.GetServices(),
 	}
 }