+++ cli/pkg/stack_mgr/stack_service.go
@@ -34,6 +34,7 @@ type StackService struct {
 	workspaceRepo workspacerepo.WorkspaceRepoIface
 	dirProcessor  util.DirProcessor
 	executor      util.Executor
+	happyConfig   *config.HappyConfig
 
 	// NOTE: creator Workspace is a workspace that creates dependent workspaces with
 	// given default values and configuration
@@ -52,23 +53,29 @@ func NewStackService() *StackService {
 		stacks:       nil,
 		dirProcessor: dirProcessor,
 		executor:     util.NewDefaultExecutor(),
+		happyConfig:  nil,
 	}
 }
 
 func (s *StackService) GetWritePath() string {
-	return fmt.Sprintf("/happy/%s/stacklist", s.backend.Conf().GetEnv())
+	return fmt.Sprintf("/happy/%s/stacklist", s.happyConfig.GetEnv())
 }
 
 func (s *StackService) GetNamespacedWritePath() string {
-	return fmt.Sprintf("/happy/%s/%s/stacklist", s.backend.Conf().App(), s.backend.Conf().GetEnv())
+	return fmt.Sprintf("/happy/%s/%s/stacklist", s.happyConfig.App(), s.happyConfig.GetEnv())
 }
 
 func (s *StackService) WithBackend(backend *backend.Backend) *StackService {
-	creatorWorkspaceName := fmt.Sprintf("env-%s", backend.Conf().GetEnv())
+	creatorWorkspaceName := fmt.Sprintf("env-%s", s.happyConfig.GetEnv())
 
 	s.creatorWorkspaceName = creatorWorkspaceName
 	s.backend = backend
 
+	return s
+}
+
+func (s *StackService) WithHappyConfig(happyConfig *config.HappyConfig) *StackService {
+	s.happyConfig = happyConfig
 	return s
 }
 
@@ -85,8 +92,8 @@ func (s *StackService) WithWorkspaceRepo(workspaceRepo workspacerepo.WorkspaceRe
 func (s *StackService) NewStackMeta(stackName string) *StackMeta {
 	// TODO: what are all these translations?
 	dataMap := map[string]string{
-		"app":      s.backend.Conf().App(),
-		"env":      s.backend.Conf().GetEnv(),
+		"app":      s.happyConfig.App(),
+		"env":      s.happyConfig.GetEnv(),
 		"instance": stackName,
 	}
 
@@ -122,7 +129,7 @@ func (s *StackService) NewStackMeta(stackName string) *StackMeta {
 }
 
 func (s *StackService) GetConfig() *config.HappyConfig {
-	return &s.backend.Conf().HappyConfig
+	return s.happyConfig
 }
 
 // Invoke a specific TFE workspace that creates/deletes TFE workspaces,
@@ -364,7 +371,7 @@ func (s *StackService) GetStack(ctx context.Context, stackName string) (*Stack, 
 
 // pre-format stack name and call workspaceRepo's GetWorkspace method
 func (s *StackService) GetStackWorkspace(ctx context.Context, stackName string) (workspacerepo.Workspace, error) {
-	workspaceName := fmt.Sprintf("%s-%s", s.backend.Conf().GetEnv(), stackName)
+	workspaceName := fmt.Sprintf("%s-%s", s.happyConfig.GetEnv(), stackName)
 
 	ws, err := s.workspaceRepo.GetWorkspace(ctx, workspaceName)
 	if err != nil {